Commiting updates from yesterday
