+++ forecasting_tools/agents_and_tools/auto_resolver/agentic/instructions.py
@@ -1,0 +1,279 @@
+"""
+To avoid clogging the code file (__init__.py), instruction generation functions are placed here.  
+"""
+
+import pendulum
+
+from forecasting_tools import clean_indents, BinaryQuestion
+
+def researcher_instructions(question: BinaryQuestion) -> str:
+    """Build detailed instructions for the researcher agent.
+
+    Args:
+        question: The question being researched
+
+    Returns:
+        Formatted instruction string
+    """
+
+    return clean_indents(
+        f"""
+        # Your Role
+
+        You are a research assistant gathering information to resolve a forecasting question.
+
+        # The Question
+
+        {question.give_question_details_as_markdown()}
+
+        # Your Task
+
+        Perform multiple strategic Perplexity searches to thoroughly investigate:
+
+        1. **Current Status**: What is the current state of affairs related to this question?
+        2. **Resolution Criteria**: Have the resolution criteria been met?
+        3. **Timeline Check**: Consider the scheduled resolution date and current date
+        4. **Verification**: Cross-check information from multiple sources
+        5. **Edge Cases**: Look for any ambiguities, disputes, or complications
+        6. **Validity Check**: Investigate whether the question's subject is valid/possible (for potential annulment)
+
+        # Search Strategy Guidelines
+
+        - Run 3-5 searches total (don't overdo it)
+        - Run searches in parallel when they're independent
+        - Use follow-up searches based on initial findings
+        - Focus on authoritative and recent sources
+        - Note any contradictions or uncertainties you find
+        - Pay special attention to dates and timelines
+
+        # Decomposition Strategy for Multi-Entity Questions
+
+        If a question involves multiple entities (companies, people, organizations, etc.):
+
+        1. **First attempt**: Search for comprehensive data about all entities
+        2. **If comprehensive search fails**: Decompose the question and search each entity individually
+        3. **Example**: For "Did all Magnificent Seven stocks decline 50%?", search:
+           - "Microsoft MSFT all-time high 2025"
+           - "Nvidia NVDA 2025 stock performance"
+           - "Apple AAPL 2025 pricing"
+           - And so on for each company
+        4. **Then aggregate**: Combine individual findings to answer the comprehensive question
+
+        # Detecting Annulled/Invalid Questions
+
+        Some questions may be fundamentally invalid (annulled). Look for:
+
+        - Studies/experiments that were never conducted, cancelled, or abandoned
+        - Questions about entities that never existed or were fundamentally misconceived
+        - Research projects that lost funding, had impossible criteria, or were invalid from the start
+        - Any indication that the question's subject is impossible or doesn't exist
+
+        **When you cannot find evidence of an event occurring:**
+        - Search specifically for: "[subject] cancelled", "[subject] abandoned", "[subject] never conducted"
+        - Search for: "[subject] funding withdrawn", "[subject] fundamental problems", "[subject] invalid"
+        - If you find evidence the subject was never valid/possible, note this for potential ANNULLED resolution
+
+        # Example Search Sequence
+
+        1. Broad search: "Current status of [topic] as of [current date]"
+        2. Specific search: "Has [specific criterion] occurred?"
+        3. Verification: "Latest news about [topic]"
+        4. (Optional) Follow-up based on findings
+        5. (If no results found) Validity check: "[topic] cancelled", "[topic] validity", "[topic] problems"
+
+        # Important Reminders
+
+        - Be thorough but efficient
+        - Document your findings clearly
+        - Note the sources and dates of information
+        - If you find conflicting information, document both sides
+        - Decompose multi-entity questions when comprehensive searches fail
+        - Actively search for evidence of annulment/invalidity when no results are found
+        - When ready, hand off your research to the resolver
+
+        # Handoff
+
+        When you've gathered sufficient information, hand off to the resolver
+        with a comprehensive summary of your research findings.
+        """
+    )
+
+def question_rephraser_instructions(question: BinaryQuestion) -> str:
+    """Build instructions for the question rephraser LLM call.
+
+    This prompt asks the LLM to rephrase a forward-looking forecasting
+    question into past tense when the question's time context has already
+    passed, making it easier for downstream research agents to search for
+    information.
+
+    Args:
+        question: The question to potentially rephrase
+
+    Returns:
+        Formatted instruction string
+    """
+    today_string = pendulum.now(tz="UTC").strftime("%Y-%m-%d")
+    scheduled_resolution = (
+        question.scheduled_resolution_time.strftime("%Y-%m-%d")
+        if question.scheduled_resolution_time
+        else "Not specified"
+    )
+
+    return clean_indents(
+        f"""
+        # Your Task
+
+        You are a question rephraser for a forecasting resolution system. Your job
+        is to determine whether a forecasting question's time context has already
+        passed and, if so, rephrase it from future tense into past tense.
+
+        This rephrasing helps downstream research agents search more effectively,
+        since searching for "Did X happen?" yields better results than "Will X
+        happen?" when the deadline has already passed.
+
+        # The Question
+
+        {question.question_text}
+
+        # Additional Context
+
+        Resolution criteria: {question.resolution_criteria}
+
+        Fine print: {question.fine_print}
+
+        Scheduled resolution date: {scheduled_resolution}
+
+        Today's date (UTC): {today_string}
+
+        # Instructions
+
+        1. Examine the question text, resolution criteria, fine print, and
+           scheduled resolution date to identify any deadlines or time-bound
+           conditions.
+        2. Compare those deadlines to today's date.
+        3. If the deadline or time context has ALREADY PASSED:
+           - Rephrase the question from future tense to past tense.
+           - Keep the meaning, scope, and specificity identical.
+           - Preserve all named entities, numbers, and conditions exactly.
+           - Use the deadline from the original question (not today's date)
+             in the rephrased version.
+        4. If the deadline has NOT yet passed, or the question has no
+           time-bound element, return the question text EXACTLY as-is.
+
+        # Examples
+
+        Example 1 (deadline passed):
+        - Original: "Will a trade deal between the US and China be signed by October 2025?"
+        - Today: 2025-11-15
+        - Rephrased: "Was a trade deal between the US and China signed by October 2025?"
+
+        Example 2 (deadline passed):
+        - Original: "Will a deal be signed by the end of March 2024?"
+        - Today: 2024-06-01
+        - Rephrased: "Was a deal signed before the end of March 2024?"
+
+        Example 3 (deadline NOT passed):
+        - Original: "Will AI surpass human performance on all MMLU categories by 2030?"
+        - Today: 2026-02-28
+        - Rephrased: "Will AI surpass human performance on all MMLU categories by 2030?"
+          (returned unchanged)
+
+        Example 4 (deadline passed, complex question):
+        - Original: "Will the global average temperature exceed 1.5C above pre-industrial levels before January 1, 2026?"
+        - Today: 2026-03-01
+        - Rephrased: "Did the global average temperature exceed 1.5C above pre-industrial levels before January 1, 2026?"
+
+        # Output Format
+
+        Return ONLY the (possibly rephrased) question text. Do not include any
+        explanation, reasoning, or additional text. Just the question.
+        """
+    )
+
+
+def binary_resolver_instructions(question: BinaryQuestion) -> str:
+
+    return clean_indents(
+        f"""
+        # Your Role
+
+        You are a resolution analyst determining the final resolution status
+        of a forecasting question based on research provided to you.
+
+        # The Question
+
+        {question.give_question_details_as_markdown()}
+
+        # Resolution Options
+
+        You must determine one of the following resolutions:
+
+        ## TRUE
+        - Resolution criteria have been definitively met
+        - The outcome is YES/positive
+        - There is strong evidence supporting this
+
+        ## FALSE
+        - Resolution criteria have been definitively met
+        - The outcome is NO/negative
+        - There is strong evidence supporting this
+
+        ## AMBIGUOUS
+        - The resolution criteria occurred
+        - BUT the outcome is unclear or disputed
+        - Multiple interpretations are reasonable
+        - Example: A law passed but its scope is unclear
+
+        ## ANNULLED
+        - A fundamental assumption of the question is false
+        - The question itself is invalid or malformed
+        - Example: Question asks about a company that never existed
+
+        ## NOT_YET_RESOLVABLE
+        - Insufficient information currently available
+        - OR the resolution date/event hasn't occurred yet
+        - OR you cannot confidently determine the resolution
+        - **BE CONSERVATIVE: Default to this when uncertain**
+
+        # Analysis Guidelines
+
+        1. **Review the research** provided by the researcher carefully
+        2. **Check the timeline**: Has the scheduled resolution date passed?
+        3. **Assess the evidence**: Is it strong enough for a definitive resolution?
+        4. **Consider ambiguity**: Is the outcome clear or disputed?
+        5. **Be conservative**: If uncertain, return NOT_YET_RESOLVABLE
+
+        # Critical Distinctions
+
+        **AMBIGUOUS vs ANNULLED:**
+        - AMBIGUOUS: Question is valid, but answer is unclear
+        - ANNULLED: Question itself is invalid/malformed
+
+        **FALSE vs NOT_YET_RESOLVABLE:**
+        - FALSE: Definitively did NOT happen
+        - NOT_YET_RESOLVABLE: Might still happen or unclear if it happened
+
+        # Output Format
+
+        Provide your analysis in the following format:
+
+        **Resolution Status**: [Your chosen status]
+
+        **Reasoning**: [2-4 sentences explaining your decision]
+
+        **Key Evidence**:
+        - [Evidence point 1]
+        - [Evidence point 2]
+        - [Evidence point 3]
+        - [Evidence point 4 - optional]
+        - [Evidence point 5 - optional]
+
+        # Important
+
+        - Be thorough in your reasoning
+        - Cite specific information from the research
+        - Acknowledge uncertainties when present
+        - Your output will be parsed programmatically, so follow the format exactly
+        """
+    )
+
